requirements.txt
asgiref==3.9.0
certifi==2025.6.15
charset-normalizer==3.4.2
Django==4.2
django-faker==0.2
django-filter==25.1
django-jalali==7.4.0
django-jalali-date==2.0.0
django-mail-templated==2.6.5
django-social-share==2.3.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.10
fake-factory==9999.9.9
Faker==37.5.3
idna==3.10
importlib_metadata==8.7.0
inflection==0.5.1
jalali_core==1.0.0
jdatetime==5.2.0
Markdown==3.8.2
packaging==25.0
pillow==11.3.0
PyJWT==2.9.0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
requests==2.32.4
sqlparse==0.5.3
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
zipp==3.23.0
